requirements.txt
Django
djangorestframework
PyJWT
python-decouple